+++ pkgs/http2/example/display_headers.dart
@@ -50,8 +50,8 @@ main(List<String> args) async {
 Future<Socket> connect(Uri uri) async {
   bool useSSL = uri.scheme == 'https';
   if (useSSL) {
-    var secureSocket = await SecureSocket
-        .connect(uri.host, uri.port, supportedProtocols: ['h2']);
+    var secureSocket = await SecureSocket.connect(uri.host, uri.port,
+        supportedProtocols: ['h2']);
     if (secureSocket.selectedProtocol != 'h2') {
       throw new Exception("Failed to negogiate http/2 via alpn. Maybe server "
           "doesn't support http/2.");