+++ pkgs/http2/lib/src/streams/stream_handler.dart
@@ -150,6 +150,10 @@ class StreamHandler extends Object with TerminatableMixin, ClosableMixin {
 
   bool get ranOutOfStreamIds => _ranOutOfStreamIds();
 
+  /// Whether it is possible to open a new stream to the remote end (e.g. based
+  /// on whether we have reached the limit of maximum concurrent open streams).
+  bool get canOpenStream => _canCreateNewStream();
+
   final ActiveStateHandler _onActiveStateChanged;
 
   StreamHandler._(
@@ -188,6 +192,12 @@ class StreamHandler extends Object with TerminatableMixin, ClosableMixin {
     _openStreams.values.toList().forEach((stream) =>
         _closeStreamAbnormally(stream, exception, propagateException: true));
     startClosing();
+  }
+
+  void forceDispatchIncomingMessages() {
+    _openStreams.forEach((int streamId, Http2StreamImpl stream) {
+      stream.incomingQueue.forceDispatchIncomingMessages();
+    });
   }
 
   Stream<TransportStream> get incomingStreams => _newStreamsC.stream;
@@ -645,8 +655,8 @@ class StreamHandler extends Object with TerminatableMixin, ClosableMixin {
 
   void _handleRstFrame(Http2StreamImpl stream, RstStreamFrame frame) {
     stream._handleTerminated(frame.errorCode);
-    var exception =
-        new StreamException(stream.id, 'Stream was terminated by peer.');
+    var exception = new StreamTransportException(
+        'Stream was terminated by peer (errorCode: ${frame.errorCode}).');
     _closeStreamAbnormally(stream, exception, propagateException: true);
   }